Varios cambios realizados. Apunto de cambiar el controlador, repositorio y servicios del documento para que me de los documentos en los los meses activos
Deseenme suerte

+++ src/main/java/com/example/demo/model/dto/DtoSaldos.java
@@ -1,5 +1,7 @@
 package com.example.demo.model.dto;
 
+import java.math.BigDecimal;
+
 import com.example.demo.model.pojos.Cuenta;
 import com.example.demo.model.pojos.SaldoCuenta;
 
@@ -16,9 +18,9 @@ public class DtoSaldos {
 	
 	private Cuenta cuenta;
 	
-	private float debito;
+	private BigDecimal debito;
 	
-	private float credito;
+	private BigDecimal credito;
 	
 	public static DtoSaldos extraerDatosSaldos(SaldoCuenta s) {
